requirements.txt
Flask
google-cloud-vision
pillow
transformers
torch
